api/pkg/gc: prune remote references as a part of the git GC
--- sturdy ---
user_id: 847dfd0c-49bf-40c5-8870-74a12fca0d60
view_id: 128badfe-5c58-487e-adfd-c4f1e4fd160f
workspace_id: ac83b73f-5568-44bc-af8b-db64c50f006d
